adding pink noise and fizing growth function

+++ src/GrowthFunction.java
@@ -6,22 +6,18 @@ public class GrowthFunction {
     }
     
     public float calculateSize(Plant plant) {
-
         float q = plant.getQ();
         float lifeSpan = plant.getLifeSpan();
         float currentAge = plant.getAge();
         float maxHeight;
 
-        //check the allometry of the plant at that point
-        if (plant.isAllometryOpen()){
+        if (plant.isAllometryOpen()) {
             maxHeight = plant.getMaxHeightOpen();
-        }else{
-            maxHeight = plant.getMaxHeightClosed(); 
+        } else {
+            maxHeight = plant.getMaxHeightClosed();
         }
 
-        float plantHeight = (2 / (1 + (Math.exp((currentAge/lifeSpan)*q))) - 1) * maxHeight;
-        // where t is the current tree currentAge in years, tm is the maximum achievable tree currentAge in years,
-        // mh is maximum achieve height in metres.
+        float plantHeight = (float) ((2.0 / (1.0 + Math.exp((currentAge / lifeSpan) * q))) - 1.0) * maxHeight;
         return plantHeight;
     }
 }
